added common code for respondent
