+++ src/main/steps/common/keep-details-private/private_details_confirmed/content.test.ts
@@ -1,0 +1,63 @@
+import languageAssertions from '../../../../../test/unit/utils/languageAssertions';
+import { FormContent } from '../../../../app/form/Form';
+import { CommonContent } from '../../../common/common.content';
+
+import { generateContent } from './content';
+
+const enContent = {
+  title: 'The court will keep your contact details private',
+  line1: 'You have told us you want to keep these contact details private',
+  address: 'Address',
+  email: 'Email',
+  line2: 'What the court will do',
+  line3:
+    'The court will hold this information securely and will not share it with anyone except Cafcass (Children and Family Court Advisory and Support Service) or Cafcass CYMRU unless it is by order of the court.',
+  continue: 'Continue',
+};
+
+const cyContent = {
+  title: 'The court will keep your contact details private',
+  line1: 'You have told us you want to keep these contact details private',
+  address: 'Address',
+  email: 'Email',
+  line2: 'What the court will do',
+  line3:
+    'The court will hold this information securely and will not share it with anyone except Cafcass (Children and Family Court Advisory and Support Service) or Cafcass CYMRU unless it is by order of the court.',
+  continue: 'Continue',
+};
+
+jest.mock('../../../../app/form/validation');
+/* eslint-disable @typescript-eslint/ban-types */
+describe('citizen-home content', () => {
+  const commonContent = { language: 'en' } as CommonContent;
+  let generatedContent;
+  let form;
+  beforeEach(() => {
+    generatedContent = generateContent(commonContent);
+    form = generatedContent.form as FormContent;
+  });
+
+  test('should return correct english content', () => {
+    expect(generatedContent.title).toEqual('The court will keep your contact details private');
+    expect(generatedContent.line1).toEqual('You have told us you want to keep these contact details private');
+    expect(generatedContent.line2).toEqual('What the court will do');
+    expect(generatedContent.line3).toEqual(
+      'The court will hold this information securely and will not share it with anyone except Cafcass (Children and Family Court Advisory and Support Service) or Cafcass CYMRU unless it is by order of the court.'
+    );
+  });
+
+  // eslint-disable-next-line jest/expect-expect
+  test('should return correct english content Data', () => {
+    languageAssertions('en', enContent, () => generateContent(commonContent));
+  });
+
+  // eslint-disable-next-line jest/expect-expect
+  test('should return correct welsh content', () => {
+    languageAssertions('cy', cyContent, () => generateContent({ ...commonContent, language: 'cy' }));
+  });
+
+  test('should contain continue button', () => {
+    expect((form.submit?.text as Function)(generatedContent)).toBe('Continue');
+  });
+});
+/* eslint-enable @typescript-eslint/ban-types */
