+++ src/main/steps/common/keep-details-private/private_details_not_confirmed/content.test.ts
@@ -1,0 +1,53 @@
+import languageAssertions from '../../../../../test/unit/utils/languageAssertions';
+import { FormContent } from '../../../../app/form/Form';
+import { CommonContent } from '../../../common/common.content';
+
+import { generateContent } from './content';
+
+const enContent = {
+  title: 'The court will not keep your contact details private',
+  line1:
+    'You have told us you do not want to keep your contact details private from the other people in this application.',
+  continue: 'Continue',
+};
+
+const cyContent = {
+  title: 'The court will not keep your contact details private',
+  line1:
+    'You have told us you do not want to keep your contact details private from the other people in this application.',
+  continue: 'Continue',
+};
+
+jest.mock('../../../../app/form/validation');
+/* eslint-disable @typescript-eslint/ban-types */
+describe('citizen-home content', () => {
+  const commonContent = { language: 'en' } as CommonContent;
+  let generatedContent;
+  let form;
+  beforeEach(() => {
+    generatedContent = generateContent(commonContent);
+    form = generatedContent.form as FormContent;
+  });
+
+  test('should return correct english content', () => {
+    expect(generatedContent.title).toEqual('The court will not keep your contact details private');
+    expect(generatedContent.line1).toEqual(
+      'You have told us you do not want to keep your contact details private from the other people in this application.'
+    );
+  });
+
+  // eslint-disable-next-line jest/expect-expect
+  test('should return correct english content Data', () => {
+    languageAssertions('en', enContent, () => generateContent(commonContent));
+  });
+
+  // eslint-disable-next-line jest/expect-expect
+  test('should return correct welsh content', () => {
+    languageAssertions('cy', cyContent, () => generateContent({ ...commonContent, language: 'cy' }));
+  });
+
+  test('should contain continue button', () => {
+    expect((form.submit?.text as Function)(generatedContent)).toBe('Continue');
+  });
+});
+/* eslint-enable @typescript-eslint/ban-types */
