+++ src/main/steps/common/keep-details-private/private_details_not_confirmed/template.njk
@@ -1,0 +1,14 @@
+{% extends "common/template.njk" %}
+
+{% block page_content %}
+    <fieldset class="govuk-fieldset">
+        <legend class="govuk-fieldset__legend govuk-fieldset__legend--xl">
+            <h1 class="govuk-fieldset__heading">
+                       {{title}} 
+                      </h1>
+        </legend>
+        <div class="govuk-!-padding-bottom-4"></div>
+        <p class="govuk-body-m">{{line1}}</p>
+
+    </fieldset>
+{% endblock %}
