+++ src/main/steps/common/keep-details-private/start_alternative/template.njk
@@ -1,0 +1,3 @@
+
+{% extends "common/template.njk" %}
+
