+++ src/main/steps/respondent/keep-details-private/private_details_confirmed/template.njk
@@ -1,24 +1,1 @@
-{% extends "common/template.njk" %}
-
-{% block page_content %}
-    <div class="govuk-form-group">
-        <fieldset class="govuk-fieldset">
-            <legend class="govuk-fieldset__legend govuk-fieldset__legend--xl">
-                <h1 class="govuk-fieldset__heading">
-                        {{title}}
-                      </h1>
-            </legend>
-            <div class="govuk-!-padding-bottom-4"></div>
-            <p class="govuk-body-m">{{line1}}:</p>
-
-            <ul class="govuk-list govuk-list--bullet">
-                <li>{{address}}</li>
-                <li>{{email}}</li>
-            </ul>
-
-            <h2 class="govuk-heading-m">{{line2}}</h2>
-            <p class="govuk-body-m">{{line3}}</p>
-
-        </fieldset>
-    </div>
-{% endblock %}
+{% extends "common/keep-details-private/private_details_confirmed/template.njk" %}
