+++ src/main/steps/respondent/keep-details-private/private_details_not_confirmed/content.ts
@@ -1,36 +1,10 @@
 import { TranslationFn } from '../../../../app/controller/GetController';
-import { FormContent } from '../../../../app/form/Form';
+import { generateContent as privateDetailsNotConfirmedGenerateContent } from '../../../common/keep-details-private/private_details_not_confirmed/content';
 
-const en = {
-  title: 'The court will not keep your contact details private',
-  line1:
-    'You have told us you do not want to keep your contact details private from the other people in this application.',
-  continue: 'Continue',
-};
-
-const cy: typeof en = {
-  title: 'The court will not keep your contact details private',
-  line1:
-    'You have told us you do not want to keep your contact details private from the other people in this application.',
-  continue: 'Continue',
-};
-
-const languages = {
-  en,
-  cy,
-};
-
-export const form: FormContent = {
-  fields: {},
-  submit: {
-    text: l => l.continue,
-  },
-};
-
+export { form } from '../../../common/keep-details-private/private_details_not_confirmed/content';
 export const generateContent: TranslationFn = content => {
-  const translations = languages[content.language];
+  const privateDetailsNotConfirmedContent = privateDetailsNotConfirmedGenerateContent(content);
   return {
-    ...translations,
-    form,
+    ...privateDetailsNotConfirmedContent,
   };
 };
