+++ src/main/steps/respondent/keep-details-private/start_alternative/template.njk
@@ -1,3 +1,1 @@
-
-{% extends "common/template.njk" %}
-
+{% extends "common/keep-details-private/start_alternative/template.njk" %}
